Refactor forum file
