+++ FoodOverflowApp/views/__init__.py
@@ -4,3 +4,5 @@ from .password_reset import *
 from .token import *
 from .forum import *
 from .profile_settings import *
+from .publications import *
+from .recipes import *
